Adds transaction addition operators
Implements `operator+` for `Transaction` and `Total` classes.
This allows for convenient merging of transactions into a total.
The implementation ensures that the Transaction is owned by a shared pointer to avoid dangling pointers.

Additionally, this commit ensures that the Transaction class and friend operators are properly exported from the Model library.
This also fixes a test where the expected signal wasn't being triggered due to incorrect assumptions of the signal being fired.